manless/controller/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
aiohttp==3.9.1
opencv-python==4.8.1.78
pyserial==3.5
pydantic==2.5.0
python-multipart==0.0.6
asyncio-mqtt==0.16.1 